requirements.txt
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
Django==5.1.7
django-cognito-jwt==0.0.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
Markdown==3.7
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
pyotp
qrcode
pillow
